+++ examples/bank/lib/bank/bank_account.ex
@@ -2,8 +2,24 @@ defmodule Bank.BankAccount do
   @behaviour Incident.Aggregate
 
   alias Bank.BankAccountState
-  alias Bank.Commands.{DepositMoney, OpenAccount, WithdrawMoney}
-  alias Bank.Events.{AccountOpened, MoneyDeposited, MoneyWithdrawn}
+
+  alias Bank.Commands.{
+    DepositMoney,
+    OpenAccount,
+    ReceiveMoney,
+    RevertMoneySent,
+    SendMoney,
+    WithdrawMoney
+  }
+
+  alias Bank.Events.{
+    AccountOpened,
+    MoneyDeposited,
+    MoneyReceived,
+    MoneySent,
+    MoneySentReverted,
+    MoneyWithdrawn
+  }
 
   @impl true
   def execute(%OpenAccount{account_number: account_number}) do
@@ -58,6 +74,61 @@ defmodule Bank.BankAccount do
   end
 
   @impl true
+  def execute(%SendMoney{aggregate_id: aggregate_id, amount: amount} = command) do
+    with %{aggregate_id: aggregate_id} = state when not is_nil(aggregate_id) <-
+           BankAccountState.get(aggregate_id),
+         true <- state.balance >= amount do
+      new_event = %MoneySent{
+        aggregate_id: aggregate_id,
+        transfer_id: command.transfer_id,
+        amount: amount,
+        version: state.version + 1
+      }
+
+      {:ok, new_event, state}
+    else
+      %{aggregate_id: nil} -> {:error, :account_not_found}
+      false -> {:error, :no_enough_balance}
+    end
+  end
+
+  @impl true
+  def execute(%ReceiveMoney{aggregate_id: aggregate_id, amount: amount} = command) do
+    case BankAccountState.get(aggregate_id) do
+      %{aggregate_id: aggregate_id} = state when not is_nil(aggregate_id) ->
+        new_event = %MoneyReceived{
+          aggregate_id: aggregate_id,
+          transfer_id: command.transfer_id,
+          amount: amount,
+          version: state.version + 1
+        }
+
+        {:ok, new_event, state}
+
+      %{aggregate_id: nil} ->
+        {:error, :account_not_found}
+    end
+  end
+
+  @impl true
+  def execute(%RevertMoneySent{aggregate_id: aggregate_id, amount: amount} = command) do
+    case BankAccountState.get(aggregate_id) do
+      %{aggregate_id: aggregate_id} = state when not is_nil(aggregate_id) ->
+        new_event = %MoneySentReverted{
+          aggregate_id: aggregate_id,
+          transfer_id: command.transfer_id,
+          amount: amount,
+          version: state.version + 1
+        }
+
+        {:ok, new_event, state}
+
+      %{aggregate_id: nil} ->
+        {:error, :account_not_found}
+    end
+  end
+
+  @impl true
   def apply(%{event_type: "AccountOpened"} = event, state) do
     %{
       state
@@ -84,6 +155,36 @@ defmodule Bank.BankAccount do
     %{
       state
       | balance: state.balance - event.event_data["amount"],
+        version: event.version,
+        updated_at: event.event_date
+    }
+  end
+
+  @impl true
+  def apply(%{event_type: "MoneySent"} = event, state) do
+    %{
+      state
+      | balance: state.balance - event.event_data["amount"],
+        version: event.version,
+        updated_at: event.event_date
+    }
+  end
+
+  @impl true
+  def apply(%{event_type: "MoneySentReverted"} = event, state) do
+    %{
+      state
+      | balance: state.balance + event.event_data["amount"],
+        version: event.version,
+        updated_at: event.event_date
+    }
+  end
+
+  @impl true
+  def apply(%{event_type: "MoneyReceived"} = event, state) do
+    %{
+      state
+      | balance: state.balance + event.event_data["amount"],
         version: event.version,
         updated_at: event.event_date
     }