+++ examples/bank/lib/bank/bank_account_command_handler.ex
@@ -1,3 +1,5 @@
 defmodule Bank.BankAccountCommandHandler do
-  use Incident.CommandHandler, aggregate: Bank.BankAccount, event_handler: Bank.EventHandler
+  use Incident.CommandHandler,
+    aggregate: Bank.BankAccount,
+    event_handler: Bank.BankAccountEventHandler
 end