+++ examples/bank/lib/bank/bank_account_event_handler.ex
@@ -1,4 +1,4 @@
-defmodule Bank.EventHandler do
+defmodule Bank.BankAccountEventHandler do
   @behaviour Incident.EventHandler
 
   alias Bank.Projections.BankAccount
@@ -27,7 +27,6 @@ defmodule Bank.EventHandler do
 
     data = %{
       aggregate_id: new_state.aggregate_id,
-      account_number: new_state.account_number,
       balance: new_state.balance,
       version: event.version,
       event_id: event.event_id,
@@ -43,7 +42,51 @@ defmodule Bank.EventHandler do
 
     data = %{
       aggregate_id: new_state.aggregate_id,
-      account_number: new_state.account_number,
+      balance: new_state.balance,
+      version: event.version,
+      event_id: event.event_id,
+      event_date: event.event_date
+    }
+
+    ProjectionStore.project(BankAccount, data)
+  end
+
+  @impl true
+  def listen(%{event_type: "MoneySent"} = event, state) do
+    new_state = Aggregate.apply(event, state)
+
+    data = %{
+      aggregate_id: new_state.aggregate_id,
+      balance: new_state.balance,
+      version: event.version,
+      event_id: event.event_id,
+      event_date: event.event_date
+    }
+
+    ProjectionStore.project(BankAccount, data)
+  end
+
+  @impl true
+  def listen(%{event_type: "MoneyReceived"} = event, state) do
+    new_state = Aggregate.apply(event, state)
+
+    data = %{
+      aggregate_id: new_state.aggregate_id,
+      balance: new_state.balance,
+      version: event.version,
+      event_id: event.event_id,
+      event_date: event.event_date
+    }
+
+    ProjectionStore.project(BankAccount, data)
+  end
+
+  @impl true
+  def listen(%{event_type: "MoneySentReverted"} = event, state) do
+    new_state = Aggregate.apply(event, state)
+
+    data = %{
+      aggregate_id: new_state.aggregate_id,
       balance: new_state.balance,
       version: event.version,
       event_id: event.event_id,