+++ examples/bank/lib/bank/events/transfer_reverted.ex
@@ -1,0 +1,9 @@
+defmodule Bank.Events.TransferReverted do
+  use Ecto.Schema
+
+  @primary_key false
+  embedded_schema do
+    field(:aggregate_id, :string)
+    field(:version, :integer)
+  end
+end