+++ examples/bank/lib/bank/transfer_command_handler.ex
@@ -1,0 +1,3 @@
+defmodule Bank.TransferCommandHandler do
+  use Incident.CommandHandler, aggregate: Bank.Transfer, event_handler: Bank.TransferEventHandler
+end